Refactor test_rhel_os.py
