+++ tests/infrastructure/instance_types/supported_os/test_rhel_os.py
@@ -22,6 +22,7 @@ from utilities.virt import (
 
 pytestmark = pytest.mark.post_upgrade
 
+TESTS_MODULE_IDENTIFIER = "TestCommonInstancetypeRhel"
 TEST_CLASS_CREATION_VALIDATION = "TestVMCreationAndValidation"
 TEST_CREATE_VM_TEST_NAME = f"{TEST_CLASS_CREATION_VALIDATION}::test_create_vm"
 TEST_START_VM_TEST_NAME = f"{TEST_CLASS_CREATION_VALIDATION}::test_start_vm"
@@ -35,7 +36,7 @@ TESTS_MIGRATE_VM = f"{TESTS_CLASS_MIGRATION}::test_migrate_vm"
 @pytest.mark.gating
 @pytest.mark.sno
 class TestVMCreationAndValidation:
-    @pytest.mark.dependency(name=TEST_CREATE_VM_TEST_NAME)
+    @pytest.mark.dependency(name=f"{TESTS_MODULE_IDENTIFIER}::TEST_CREATE_VM_TEST_NAME")
     @pytest.mark.polarion("CNV-11710")
     def test_create_vm(self, golden_image_vm_with_instance_type, instance_type_rhel_os_matrix__module__):
         golden_image_vm_with_instance_type.create(wait=True)
@@ -43,24 +44,27 @@ class TestVMCreationAndValidation:
         assert golden_image_vm_with_instance_type.instance.spec.instancetype.name == os_param_dict[INSTANCE_TYPE_STR]
         assert golden_image_vm_with_instance_type.instance.spec.preference.name == os_param_dict[PREFERENCE_STR]
 
-    @pytest.mark.dependency(name=TEST_START_VM_TEST_NAME, depends=[TEST_CREATE_VM_TEST_NAME])
+    @pytest.mark.dependency(
+        name=f"{TESTS_MODULE_IDENTIFIER}::TEST_START_VM_TEST_NAME",
+        depends=[f"{TESTS_MODULE_IDENTIFIER}::TEST_CREATE_VM_TEST_NAME"],
+    )
     @pytest.mark.polarion("CNV-11711")
     def test_start_vm(self, golden_image_vm_with_instance_type):
         running_vm(vm=golden_image_vm_with_instance_type)
 
-    @pytest.mark.dependency(depends=[TEST_START_VM_TEST_NAME])
+    @pytest.mark.dependency(depends=[f"{TESTS_MODULE_IDENTIFIER}::TEST_START_VM_TEST_NAME"])
     @pytest.mark.polarion("CNV-11712")
     def test_vm_console(self, golden_image_vm_with_instance_type):
         wait_for_console(vm=golden_image_vm_with_instance_type)
 
-    @pytest.mark.dependency(depends=[TEST_START_VM_TEST_NAME])
+    @pytest.mark.dependency(depends=[f"{TESTS_MODULE_IDENTIFIER}::TEST_START_VM_TEST_NAME"])
     @pytest.mark.polarion("CNV-11829")
     def test_expose_ssh(self, golden_image_vm_with_instance_type):
         assert golden_image_vm_with_instance_type.ssh_exec.executor().is_connective(tcp_timeout=120), (
             "Failed to login via SSH"
         )
 
-    @pytest.mark.dependency(depends=[TEST_START_VM_TEST_NAME])
+    @pytest.mark.dependency(depends=[f"{TESTS_MODULE_IDENTIFIER}::TEST_START_VM_TEST_NAME"])
     @pytest.mark.polarion("CNV-11713")
     def test_vmi_guest_agent_exists(self, golden_image_vm_with_instance_type):
         assert check_qemu_guest_agent_installed(ssh_exec=golden_image_vm_with_instance_type.ssh_exec), (
@@ -71,27 +75,27 @@ class TestVMCreationAndValidation:
 @pytest.mark.usefixtures("skip_if_rhel8")
 @pytest.mark.sno
 class TestVMFeatures:
-    @pytest.mark.dependency(depends=[TEST_START_VM_TEST_NAME])
+    @pytest.mark.dependency(depends=[f"{TESTS_MODULE_IDENTIFIER}::TEST_START_VM_TEST_NAME"])
     @pytest.mark.polarion("CNV-11830")
     def test_system_boot_mode(self, golden_image_vm_with_instance_type):
         assert_linux_efi(vm=golden_image_vm_with_instance_type)
 
     @pytest.mark.polarion("CNV-11831")
-    @pytest.mark.dependency(depends=[TEST_START_VM_TEST_NAME])
+    @pytest.mark.dependency(depends=[f"{TESTS_MODULE_IDENTIFIER}::TEST_START_VM_TEST_NAME"])
     def test_efi_secureboot_enabled_initial_boot(self, golden_image_vm_with_instance_type):
         assert_secure_boot_dmesg(vm=golden_image_vm_with_instance_type)
 
-    @pytest.mark.dependency(depends=[TEST_START_VM_TEST_NAME])
+    @pytest.mark.dependency(depends=[f"{TESTS_MODULE_IDENTIFIER}::TEST_START_VM_TEST_NAME"])
     @pytest.mark.polarion("CNV-11832")
     def test_efi_secureboot_enabled_guest_os(self, golden_image_vm_with_instance_type):
         assert_secure_boot_mokutil_status(vm=golden_image_vm_with_instance_type)
 
-    @pytest.mark.dependency(depends=[TEST_START_VM_TEST_NAME])
+    @pytest.mark.dependency(depends=[f"{TESTS_MODULE_IDENTIFIER}::TEST_START_VM_TEST_NAME"])
     @pytest.mark.polarion("CNV-11833")
     def test_efi_secureboot_enabled_lockdown_state(self, golden_image_vm_with_instance_type):
         assert_kernel_lockdown_mode(vm=golden_image_vm_with_instance_type)
 
-    @pytest.mark.dependency(depends=[TEST_START_VM_TEST_NAME])
+    @pytest.mark.dependency(depends=[f"{TESTS_MODULE_IDENTIFIER}::TEST_START_VM_TEST_NAME"])
     @pytest.mark.polarion("CNV-11834")
     def test_efi_secureboot_disabled_and_enabled(
         self,
@@ -110,7 +114,7 @@ class TestVMFeatures:
         # Re-enable Secure Boot
         _update_and_verify_secure_boot(vm=vm, secure_boot_value=True)
 
-    @pytest.mark.dependency(depends=[TEST_START_VM_TEST_NAME])
+    @pytest.mark.dependency(depends=[f"{TESTS_MODULE_IDENTIFIER}::TEST_START_VM_TEST_NAME"])
     @pytest.mark.polarion("CNV-11835")
     def test_vm_smbios_default(self, smbios_from_kubevirt_config, golden_image_vm_with_instance_type):
         check_vm_xml_smbios(vm=golden_image_vm_with_instance_type, cm_values=smbios_from_kubevirt_config)
@@ -119,18 +123,21 @@ class TestVMFeatures:
 @pytest.mark.arm64
 class TestVMMigrationAndState:
     @pytest.mark.polarion("CNV-11714")
-    @pytest.mark.dependency(name=TESTS_MIGRATE_VM, depends=[TEST_START_VM_TEST_NAME])
+    @pytest.mark.dependency(
+        name=f"{TESTS_MODULE_IDENTIFIER}::TESTS_MIGRATE_VM",
+        depends=[f"{TESTS_MODULE_IDENTIFIER}::TEST_START_VM_TEST_NAME"],
+    )
     def test_migrate_vm(self, skip_access_mode_rwo_scope_class, golden_image_vm_with_instance_type):
         migrate_vm_and_verify(vm=golden_image_vm_with_instance_type, check_ssh_connectivity=True)
         validate_libvirt_persistent_domain(vm=golden_image_vm_with_instance_type)
 
     @pytest.mark.polarion("CNV-11836")
-    @pytest.mark.dependency(depends=[TESTS_MIGRATE_VM])
+    @pytest.mark.dependency(depends=[f"{TESTS_MODULE_IDENTIFIER}::TESTS_MIGRATE_VM"])
     def test_pause_unpause_vm(self, golden_image_vm_with_instance_type):
         validate_pause_optional_migrate_unpause_linux_vm(vm=golden_image_vm_with_instance_type)
 
     @pytest.mark.polarion("CNV-11837")
-    @pytest.mark.dependency(depends=[TESTS_MIGRATE_VM])
+    @pytest.mark.dependency(depends=[f"{TESTS_MODULE_IDENTIFIER}::TESTS_MIGRATE_VM"])
     def test_pause_unpause_after_migrate(self, golden_image_vm_with_instance_type, ping_process_in_rhel_os):
         validate_pause_optional_migrate_unpause_linux_vm(
             vm=golden_image_vm_with_instance_type,
@@ -138,7 +145,7 @@ class TestVMMigrationAndState:
         )
 
     @pytest.mark.polarion("CNV-11838")
-    @pytest.mark.dependency(depends=[TESTS_MIGRATE_VM])
+    @pytest.mark.dependency(depends=[f"{TESTS_MODULE_IDENTIFIER}::TESTS_MIGRATE_VM"])
     def test_verify_virtctl_guest_agent_data_after_migrate(self, golden_image_vm_with_instance_type):
         assert validate_virtctl_guest_agent_data_over_time(vm=golden_image_vm_with_instance_type), (
             "Guest agent stopped responding"
@@ -151,7 +158,7 @@ class TestVMMigrationAndState:
 @pytest.mark.sno
 @pytest.mark.order(-1)
 class TestVMDeletion:
-    @pytest.mark.dependency(depends=[TEST_CREATE_VM_TEST_NAME])
+    @pytest.mark.dependency(depends=[f"{TESTS_MODULE_IDENTIFIER}::TEST_CREATE_VM_TEST_NAME"])
     @pytest.mark.polarion("CNV-11715")
     def test_vm_deletion(self, golden_image_vm_with_instance_type):
         golden_image_vm_with_instance_type.delete(wait=True)
